# openapi.yaml
openapi: "3.0.0"
info:
  title: "TASK APP - SPRING BOOT"
  version: "1.0.0"
servers:
  - url: "http://localhost:{port}"
    description: "Localhost"
    variables:
      port:
        enum:
          - "8080"
        default: "8080"
components:
  schemas:
    requestBodyOrganizationCreate:
      type: object
      properties:
        name:
          type: string
          description: "Nome da organização"
        active:
          type: string
          enum:
            - Y
            - N
    requestBodyOrganizationUpdate:
      type: object
      properties:
        id:
          type: string
          description: "Id da organização"
          format: uuid
        name:
          type: string
          description: "Nome da organização"
        active:
          type: string
          enum:
            - Y
            - N
    organization:
      type: object
      properties:
        id:
          type: string
          description: "Id da organização"
          format: uuid
        name:
          type: string
          description: "Nome da organização"
        active:
          type: string
          enum:
            - Y
            - N
paths:
  /organization:
    post:
      description: "Cria uma nova organização"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/requestBodyOrganizationCreate"
      responses:
        201:
          description: "Organização criada com sucesso"
        400:
          description: "Erro na requisição"
        500:
          description: "Erro interno"
    put:
      description: "Editar uma organização"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/requestBodyOrganizationUpdate"
      responses:
        201:
          description: "Organização editada com sucesso"
        400:
          description: "Erro na requisição"
        500:
          description: "Erro interno"
  /organization/{id}:
    get:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        201:
          description: Organização obtida com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/organization"
        404:
          description: "Organização não encontrada"